Sekoia.io-alertEventsThreshold/pyproject.toml
[tool.poetry]
name = "Automation Module for Sekoia.io"
description = ""
version = "0.0"
authors = []
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
requests = "*"
sekoia-automation-sdk = "^1.18.0"
kafka-python = "*"
tenacity = "*"
ujson = "*"
websocket-client = "*"
gevent = "*"
orjson = "*"
aiohttp = "^3.13.2"
aiolimiter = "^1.1.0"


[tool.poetry.group.dev.dependencies]
pytest = "*"
pytest-cov = "*"
requests = "*"
requests-mock = "*"
aioresponses = "^0.7.6"
aiolimiter = "^1.1.0"
pytest-asyncio = "^1.3.0"
aiohttp = "^3.13.2"

[build-system]
requires = [ "poetry-core>=1.0.0",]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 119
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 119

[tool.coverage.run]
omit = [
    "tests/*",
    "main.py",
]

[tool.ruff]
target-version = "py311"
line-length = 119
fix = true

# ============================================================================
# PYTEST CONFIGURATION - Added to fix asyncio warnings
# ============================================================================
[tool.pytest.ini_options]
# Fix pytest-asyncio deprecation warning
asyncio_default_fixture_loop_scope = "function"
asyncio_mode = "auto"

# Standard pytest settings
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]